requirements.txt
flask==3.0.0
python-dotenv==1.0.0
notion-client==2.2.1
gunicorn==21.2.0
